[Refactor]: 뭉쳐져 있던 상수 변수를 파일별로 분리
