+++ src/App.js
@@ -1,10 +1,5 @@
 import { Console } from '@woowacourse/mission-utils';
-import {
-  INPUT_MESSAGE,
-  ERROR_MESSAGE,
-  CONDITION,
-  RACE_MESSAGE,
-} from './constant/constants.js';
+import { CONDITION, ERROR, MESSAGE } from './constants/index.js';
 import Car from './Car.js';
 import Random from './Random.js';
 import Refree from './Referee.js';
@@ -13,11 +8,11 @@ class App {
   #cars;
 
   async play() {
-    const carNames = await Console.readLineAsync(INPUT_MESSAGE.CAR_NAME);
+    const carNames = await Console.readLineAsync(MESSAGE.CAR_NAME);
     const carNameArray = carNames.split(',');
     App.#validateNameLength(carNameArray);
 
-    const tryCount = await Console.readLineAsync(INPUT_MESSAGE.TRY_COUNT);
+    const tryCount = await Console.readLineAsync(MESSAGE.TRY_COUNT);
     App.#validateTryCount(tryCount);
 
     this.#cars = carNameArray.map((carName) => new Car(carName));
@@ -26,12 +21,12 @@ class App {
   }
 
   #printRaceResult(count) {
-    Console.print(RACE_MESSAGE.RESULT);
+    Console.print(MESSAGE.RESULT);
 
     for (let i = 0; i < count; i++) {
       this.#cars.forEach((car) => car.race(Random.createRandomNumber()));
       this.#cars.forEach((car) =>
-        Console.print(RACE_MESSAGE.RACE(car.getName(), car.getForwardCount()))
+        Console.print(MESSAGE.RACE(car.getName(), car.getForwardCount()))
       );
       Console.print('');
     }
@@ -39,18 +34,18 @@ class App {
 
   #printWinners() {
     const winners = Refree.judge(this.#cars);
-    Console.print(RACE_MESSAGE.WINNERS(winners));
+    Console.print(MESSAGE.WINNERS(winners));
   }
 
   static #validateNameLength(carNameArray) {
     if (CONDITION.NAME_LENGTH(carNameArray)) {
-      throw new Error(ERROR_MESSAGE.NAME_LENGTH);
+      throw new Error(ERROR.NAME_LENGTH);
     }
   }
 
   static #validateTryCount(count) {
     if (CONDITION.TRY_COUNT(Number(count))) {
-      throw new Error(ERROR_MESSAGE.TRY_COUNT);
+      throw new Error(ERROR.TRY_COUNT);
     }
   }
 }
