+++ src/constants/error.js
@@ -1,0 +1,6 @@
+const ERROR = Object.freeze({
+  NAME_LENGTH: '[ERROR] 이름은 5자 이하만 가능합니다.',
+  TRY_COUNT: '[ERROR] 시도할 횟수는 양의 정수여야 합니다.',
+});
+
+export default ERROR;
